# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axgit API",
    "description": "Read-only JSON API over bare Git repositories, plus Smart HTTP (fetch/clone) and an Atom feed.\n\nThere are no mutation endpoints and no authentication: writes happen over SSH on the git-server container, never here.\n\nThis document describes shapes, parameters and status codes. The semantic rules it cannot express — truncation caps, longest-ref matching for the `{ref}`/`{path}` split, merge-commit log simplification, and the exact conditions under which a field is `null` — live in `docs/API.md`, which remains the normative contract.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/repos/{repo}/blame/{ref}/{path}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Per-line attribution",
        "description": "`{ref}`/`{path}` split as for the tree endpoint. Rename tracking is not\nperformed — only moves within the file are attributed, per libgit2\ndefaults.",
        "operationId": "get_blame",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "path",
            "in": "path",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "src/main.rs"
          }
        ],
        "responses": {
          "200": {
            "description": "Line ranges covering the whole file. Immutable caching only when `{ref}` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlameInfo"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "400": {
            "description": "`invalid_param` — `.`, `..`, or an empty path segment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`, `path_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repo}/blob/{ref}/{path}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "File metadata and content",
        "description": "`{ref}`/`{path}` split as for the tree endpoint. Binary or over-1 MiB files\nreport `content: null` — fetch the raw endpoint instead.",
        "operationId": "get_blob",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "path",
            "in": "path",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "README.md"
          }
        ],
        "responses": {
          "200": {
            "description": "Blob metadata and inline content. Immutable caching only when `{ref}` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobInfo"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "400": {
            "description": "`invalid_param` — `.`, `..`, or an empty path segment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`, `path_not_found` (missing, or a directory/submodule)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repos/{repo}/raw/{ref}/{path}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Raw file contents",
        "description": "No size cap. Not routed through the response cache — raw bodies are\nunbounded binaries that would crowd out the JSON entries. Non-sha requests\nstill carry an `ETag` so a 304 saves the transfer (the bytes are re-read\neither way). `{ref}`/`{path}` split as for the tree endpoint.",
        "operationId": "get_raw",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "path",
            "in": "path",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "README.md"
          }
        ],
        "responses": {
          "200": {
            "description": "File bytes. `Content-Type` is guessed from the extension, falling back to `text/plain; charset=utf-8` or `application/octet-stream`.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string"
                },
                "description": "Always `nosniff` — repository contents are untrusted"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "400": {
            "description": "`invalid_param` — `.`, `..`, or an empty path segment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`, `path_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repo}/readme": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "README lookup",
        "description": "Searches the root tree for `README.md` → `README.rst` → `README.txt` →\n`README`, matched case-insensitively. Symlinks and binary or over-1 MiB\ncandidates are skipped. Rendering is the frontend's job.",
        "operationId": "get_readme",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch, tag, or commit sha; HEAD when absent.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "main"
          }
        ],
        "responses": {
          "200": {
            "description": "The first usable README. Immutable caching only when `ref` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadmeInfo"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found` (including an empty repository), `path_not_found` (no README)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repos/{repo}/tree/{ref}/{path}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Directory listing",
        "description": "The `{ref}`/`{path}` boundary is resolved per request by longest-ref\nmatching (`repo::resolve::resolve_ref_path`), because branch and tag names\nmay contain `/`; with no match the first segment is taken as the ref.\nOmit `{path}` for the root tree.",
        "operationId": "get_tree",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "path",
            "in": "path",
            "description": "Directory path; omit (empty) for the root tree",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "src"
          }
        ],
        "responses": {
          "200": {
            "description": "Directory entries, trees first then by name. Immutable caching only when `{ref}` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeListing"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "400": {
            "description": "`invalid_param` — `.`, `..`, or an empty path segment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`, `path_not_found` (missing or not a directory)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BlameInfo": {
        "type": "object",
        "description": "Response of the blame endpoint (docs/API.md).",
        "required": [
          "sha",
          "path",
          "binary",
          "too_large",
          "lines",
          "ranges"
        ],
        "properties": {
          "sha": {
            "type": "string",
            "description": "Resolved commit sha the file was blamed at."
          },
          "path": {
            "type": "string"
          },
          "binary": {
            "type": "boolean",
            "description": "libgit2's NUL heuristic, or content that is not valid UTF-8."
          },
          "too_large": {
            "type": "boolean",
            "description": "`true` past the same 1 MiB cap the blob endpoint applies."
          },
          "lines": {
            "type": "integer",
            "description": "Total line count; 0 for `binary`, `too_large`, or empty files.",
            "minimum": 0
          },
          "ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlameRange"
            },
            "description": "`start_line` ascending, covering the whole file with no gaps.\nEmpty for `binary`, `too_large`, or empty files."
          }
        }
      },
      "BlameRange": {
        "type": "object",
        "description": "One contiguous run of lines attributed to the same commit (docs/API.md).",
        "required": [
          "start_line",
          "line_count",
          "sha",
          "summary",
          "author",
          "authored_at"
        ],
        "properties": {
          "start_line": {
            "type": "integer",
            "description": "1-based, inclusive.",
            "minimum": 0
          },
          "line_count": {
            "type": "integer",
            "minimum": 0
          },
          "sha": {
            "type": "string",
            "description": "Sha of the commit the lines are attributed to."
          },
          "summary": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` for non-utf8 commit messages."
          },
          "author": {
            "$ref": "#/components/schemas/CommitAuthor"
          },
          "authored_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BlobInfo": {
        "type": "object",
        "description": "Response of the blob endpoint (docs/API.md).",
        "required": [
          "sha",
          "path",
          "mode",
          "size",
          "binary",
          "too_large",
          "content"
        ],
        "properties": {
          "sha": {
            "type": "string",
            "description": "Resolved commit sha the blob was read from."
          },
          "path": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "description": "Octal file mode, e.g. `\"100644\"` (`\"120000\"` for symlinks).",
            "example": "100644"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Full size in bytes, regardless of `too_large`.",
            "minimum": 0
          },
          "binary": {
            "type": "boolean",
            "description": "libgit2's NUL heuristic, or content that is not valid UTF-8."
          },
          "too_large": {
            "type": "boolean",
            "description": "`true` past the 1 MiB inline-content cap; fetch the raw endpoint instead."
          },
          "content": {
            "type": [
              "string",
              "null"
            ],
            "description": "UTF-8 content; `None` when `binary` or `too_large`. For a symlink this\nis the link target path."
          }
        }
      },
...
      "CommitAuthor": {
        "type": "object",
        "description": "Commit author of `GET /api/v1/repos/{repo}/commits` (docs/API.md).\nThe raw email is never exposed; `email_hash` seeds locally generated avatars.",
        "required": [
          "name",
          "email_hash"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email_hash": {
            "type": "string",
            "description": "sha256 hex of the trimmed, lowercased author email.",
            "example": "b642b4217b34b1e8d3bd915fc65c4452"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "One of `repo_not_found`, `ref_not_found`, `path_not_found`,\n`invalid_param`, `read_only`, `internal`.",
            "example": "repo_not_found"
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Internal errors report a generic message; the\ncause only goes to the server log.",
            "example": "repository 'foo' not found"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope shared by every failing endpoint (docs/API.md).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "ReadmeFormat": {
        "type": "string",
        "description": "How the frontend should render the README. Only `markdown` is converted to\nHTML; the rest are shown as plain text (docs/DECISIONS.md #11).",
        "enum": [
          "markdown",
          "rst",
          "plain"
        ]
      },
      "ReadmeInfo": {
        "type": "object",
        "required": [
          "path",
          "format",
          "content"
        ],
        "properties": {
          "path": {
            "type": "string",
            "description": "Actual file name as it appears in the tree (case may differ).",
            "example": "README.md"
          },
          "format": {
            "$ref": "#/components/schemas/ReadmeFormat"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "TreeEntryInfo": {
        "type": "object",
        "description": "Tree entry of the tree endpoint (docs/API.md).",
        "required": [
          "name",
          "type",
          "mode",
          "size"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EntryKind"
          },
          "mode": {
            "type": "string",
            "description": "Octal file mode, e.g. `\"100644\"`.",
            "example": "100644"
          },
          "size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Object size in bytes; blobs only, `None` otherwise.",
            "minimum": 0
          }
        }
      },
      "TreeListing": {
        "type": "object",
        "required": [
          "sha",
          "path",
          "entries"
        ],
        "properties": {
          "sha": {
            "type": "string",
            "description": "Resolved commit sha the listing was taken from."
          },
          "path": {
            "type": "string",
            "description": "Requested directory path; empty string for the root tree."
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TreeEntryInfo"
            },
            "description": "Trees first, then by name ascending."
          }
        }
      }
    }
  },
...
}